NSFS | NC | Refactor validation (partial)
